requirements.txt

flask-socketio
flask_restful
flask_jwt_extended
sqlalchemy_serializer
Flask-SQLAlchemy
Flask>=2.3.2
Flask-SQLAlchemy>=3.1.1
Flask-Cors>=3.0.10
Werkzeug>=2.3.8
flask-swagger-ui>=4.11.1
flask-login>=0.6.3
pytest>=7.4.3
python-dotenv>=1.0.1
pymysql
cryptography

